backend/requirements.txt
asgiref==3.3.1
Django==3.0.11
django-extensions==3.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
factory-boy==3.1.0
Faker==5.0.1
Pillow==8.0.1
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
